### 2025-06-25 #8483 @editorialbot
[PRE REVIEW]: jaxDecomp: JAX Library for 3D Domain Decomposition and Parallel FFTs
**Submitting author:** @ASKabalan (<a href="http://orcid.org/0009-0001-6501-4564">Wassim KABALAN</a>)
**Repository:** https://github.com/DifferentiableUniverseInitiative/jaxDecomp
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.7
