### 2025-06-25 #8483 @danielskatz
👋 @ASKabalan - thanks for your submission.

Your paper is far too long. JOSS expects papers of up to 1000 words, more or less.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right or after making changes to the .md file to make a new PDF.  You can also use the command `@editorialbot check repository` to check a number of things, including the paper's word count.

editorialbot commands need to be the first entry in a new comment.
