### 2025-06-25 8bebbaa
Trimming word count in paper
