+++ joss-paper/paper.md
@@ -91,59 +91,7 @@ In pencil decomposition, the 3D FFT is computed via three sequential 1D FFTs, ea
 | Transpose X to Y | $\frac{Y}{P_x} \times \frac{Z}{P_y} \times X$| $\frac{Z}{P_y} \times \frac{X}{P_x} \times Y$ |
 | Transpose Y to Z | $\frac{Z}{P_y} \times \frac{X}{P_x} \times Y$| $\frac{X}{P_x} \times \frac{Y}{P_y} \times Z$ |
 
-#### Slab Decomposition
-
-For 1D decomposition (slabs), we need to perform one 1D FFT and one 2D FFT. 2D FFTs present additional challenges because both the fastest and second-fastest axes must remain undistributed.
-
-For example, consider a $(P_y, P_z)$ decomposition with $P_z = 1$:
-
-| Step            | Decomposition                     | FFT Feasibility                 |
-|-----------------|-----------------------------------|---------------------------------|
-| Initial         | $X \times \frac{Y}{P_y} \times Z$  | only 1D FFT on Z        |
-| Transpose Z to Y| $\frac{Z}{P_y} \times X \times Y$  | 2D FFT on YX            |
-
-This is the case for the YZ slab, where the transformation sequence enables the application of a 2D FFT on the YX plane:
-
-2D FFT($YX$) $\rightarrow$ TransposeYtoX $\rightarrow$ 1D FFT($X$)
-
-For the other decomposition, with $P_y = 1$ and $P_x = n$ where $n$ is the number of GPUs :
-
-| Step            | Decomposition                    | FFT Feasibility                    |
-|-----------------|----------------------------------|------------------------------------|
-| Initial         | $\frac{X}{P_x} \times Y \times Z$| 1D FFT on Z or 2D FFT on YX |
-| Transpose Z to Y| $Z \times \frac{X}{P_x} \times Y$| 1D FFT on Y                 |
-
-Since `cuDecomp` does not support a direct transposition from a Z pencil to an X pencil, we achieve this by applying a coordinate transformation that effectively reinterprets the axes from XYZ to YZX—enabling slab decomposition in a single transposition step.
-
-
-#### Slab Decomposition with Coordinate Transformation
-
-| Step            | Decomposition                     | Transformed Coordinates                   | Coordinate Step  | FFT Feasibility              |
-|-----------------|-----------------------------------|-----------------------------------|------------------|------------------------------|
-| Initial         | $\frac{Z}{P_x} \times X \times Y$ | $\frac{X}{P_x} \times Y \times Z$ | –                | 2D FFT on ZY                 |
-| Transpose Y to Z| $X \times \frac{Y}{P_x} \times Z$ | $Y \times \frac{Z}{P_x} \times X$ | Transpose Z to X | 1D (I)FFT on the last axis X |
-| Transpose Z to Y| $\frac{Z}{P_x} \times X \times Y$ | $\frac{X}{P_x} \times Y \times Z$ | Transpose X to Z | 2D IFFT on ZY                |
-
-This approach ensures that slab decomposition can be achieved in a single transposition step, enhancing computational efficiency.
-
-
-2D FFT($ZY$) $\rightarrow$ TransposeZtoX $\rightarrow$ 1D FFT($X$)
-
-### Non-Contiguous Global Transpose
-
[…]
