### 2025-06-25 5177f89
Continue Trimming paper (#77)
