### 2025-06-25 a6c0d94
Trimming some more
