### 2025-06-25 59450ac
Even more trimming
