### 2025-06-25 2328d60
More trimming
