+++ joss-paper/paper.md
@@ -45,7 +45,7 @@ To fill this gap, we present `jaxDecomp`, a fully differentiable JAX library for
 
 For numerical simulations on HPC systems, a distributed, easy-to-use, and differentiable FFT is essential for achieving peak performance and scalability. While JAX now provides native distributed FFT support, this was introduced only very recently and lacks the specialized HPC features required by many applications. There is a pressing need for a solution that provides not only distributed FFTs but also halo exchanges, optimized communication backends, and seamless integration with existing cluster infrastructure.
 
-In scientific applications such as cosmological particle mesh (PM) simulations, specialized frameworks like `FlowPM` [@FlowPM] built on `mesh-TensorFlow` [@TF-MESH] or JAX-based codes like `pmwd` [@pmwd] often struggle to scale beyond single-node memory limits or rely on manual distribution strategies. These challenges highlight the need for a scalable, high-performance approach to distributed FFTs that remains differentiable for advanced algorithms (like Hamiltonian Monte Carlo [@HMC] or the No-U-Turn Sampler (NUTS) [@NUTS]).
+In scientific applications such as cosmological particle mesh (PM) simulations, specialized frameworks like `FlowPM` \[@FlowPM] built on `mesh-TensorFlow` \[@TF-MESH] or JAX-based codes like `pmwd` \[@pmwd] often struggle to scale beyond single-node memory limits or rely on manual distribution strategies. These challenges highlight the need for a scalable, high-performance approach to distributed FFTs that remains differentiable for advanced algorithms (like Hamiltonian Monte Carlo \[@HMC] or the No-U-Turn Sampler (NUTS) \[@NUTS]).
 
 
 # Implementation
