+++ joss-paper/paper.bib
@@ -26,6 +26,7 @@
    archivePrefix={arXiv},
    primaryClass={astro-ph.IM},
    url={https://arxiv.org/abs/2211.09958},
+   DOI={10.48550/arXiv.2211.09958},
 }
 
 @software{JAX,
@@ -44,6 +45,7 @@
    archivePrefix={arXiv},
    primaryClass={stat.CO},
    url={https://arxiv.org/abs/1111.4246},
+   DOI={10.48550/arXiv.1111.4246},
 }
 
 @book{HMC,
@@ -101,6 +103,8 @@
    archivePrefix={arXiv},
    primaryClass={cs.LG},
    url={https://arxiv.org/abs/1811.02084},
+   DOI={10.48550/arXiv.1811.02084},
+
 }
 
 
