### 2025-06-25 f1c1d1f
More trimming (#79)
* Continue Trimming paper

* Trimming some more

* Even more trimming

* More trimming

* fix citings
