### 2025-06-25 4845df5
DOI and centering figures
