+++ joss-paper/paper.bib
@@ -93,6 +93,7 @@
    archivePrefix={arXiv},
    primaryClass={astro-ph.CO},
    url={https://arxiv.org/abs/2010.11847},
+   DOI={10.48550/arXiv.2010.11847},
 }
 
 @misc{TF-MESH,
@@ -116,6 +117,7 @@
       archivePrefix={arXiv},
       primaryClass={physics.comp-ph},
       url={https://arxiv.org/abs/1912.04232},
+      DOI={10.48550/arXiv.1912.04232},
 }
 
 @article{NUMPY,
