### 2025-06-25 #8483 @ASKabalan
👋 @danielskatz
Thank you for your review.
I have trimmed the paper to ~1000 words .. most of the technical documention is still available [here](https://jaxdecomp.readthedocs.io/en/latest/index.html)
And I also fixed the DOI is the bib file

Thank you for your time.
