### 2025-06-26 #8483 @danielskatz
@ASKabalan - I note that your references aren't quite in the right style, meaning how they are mentioned in your .md file - please see the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html)
