### 2025-07-17 #8483 @lamBOOO
Hi @diehlpk. Unfortunately, I don't have that much time at the moment and therefore have to decline. 😅
