### 2025-09-03 DifferentiableUniverseInitiative/jaxDecomp#83 @MoraruMaxim
JOSS paper scaling plots
The current weak scaling plots show performance with a fixed number of GPUs while increasing problem size. To actually highlight the weak scaling performance, the workload per GPU (or node) should stay constant while increasing the number of GPUs (in a single plot). Could you update the plots to reflect this?

Additionally, could you please add performance efficiency or the “ideal” time line? Right now it’s hard to see how well the results scale without a baseline for comparison.
