### 2025-09-03 #8852 @MoraruMaxim
@ASKabalan, could you please provide the actual numbers obtained for the weak and strong scaling experiments? Also, can you confirm if the problem used for those results is the same as the one in the README under the Usage section (the 3D FFT)? I would like to check if I can reproduce the performance results.
