### 2025-09-09 #8852 @ASKabalan
Please find CSVs here and a notebook for plotting using [jax-hpc-profiler](https://github.com/ASKabalan/jax-hpc-profiler)

[PERF.tar.gz](https://github.com/user-attachments/files/22235004/PERF.tar.gz)

Just a heads-up .. this perfs are from last year, but I think that they should still be very representative of the perfs

You asked another question about the baseline .. I would assume that it is a (nlogn) line
Would you suggest adding it to the plots?

Thank you
