+++ TODO.md
@@ -1,9 +1,0 @@
-Reduce Abstract => Statement of need  / SUmmary
-
-Make DOCS with
-
-Installation instruction
-Quick example
-Choosing correct PDIMS
-
-Appendix goes to github pages
