+++ src/jaxdecomp/_src/cudecomp/fft.py
@@ -312,6 +312,23 @@ class FFTPrimitive(BasePrimitive):
         return NamedSharding(input_mesh, P(*transposed_specs))
 
     @staticmethod
+    def sharding_rule_producer(
+        fft_type: FftType,
+        adjoint: bool,
+        mesh: Mesh,
+        arg_infos: tuple[ShapeDtypeStruct],
+        result_infos: tuple[ShapedArray],
+    ) -> NamedSharding:
+        
+        del adjoint, result_infos
+
+        spec = ("i" , "j", "k") # einsum spec for shardy
+        transposed_specs = get_output_specs(fft_type, spec, mesh, 'cudecomp')
+        einsum_in = ' '.join(spec)
+        einsum_out = ' '.join(transposed_specs)
+        return f'{einsum_in}->{einsum_out}'
+
+    @staticmethod
     def partition(
         fft_type: FftType,
         adjoint: bool,

[…]
