### 2025-09-09 2f7b0de
skip shardy for jax 0.6.2
