### 2025-09-10 ee94301
format and add docstrings
