### 2025-09-10 ae4d162
Add README section about Shardy
