### 2025-09-10 6f6bdb0
Quick fix for cuDecomp lowering
