+++ src/jaxdecomp/_src/cudecomp/fft.py
@@ -343,7 +343,7 @@ class FFTPrimitive(BasePrimitive):
         del adjoint, result_infos
 
         spec = ('i', 'j', 'k')  # einsum spec for shardy
-        transposed_specs: tuple(str) = get_output_specs(fft_type, spec, mesh, 'cudecomp')  # type: ignore
+        transposed_specs: tuple[str, ...] = get_output_specs(fft_type, spec, mesh, 'cudecomp')  # type: ignore
         einsum_in = ' '.join(spec)
         einsum_out = ' '.join(transposed_specs)
         return f'{einsum_in}->{einsum_out}'
