### 2025-09-10 455d775
Implement partitionning using shardy (#86)
* Implement partitionning using shardy

* Format

* skip shardy for jax 0.6.2

* format and add docstrings

* Add README section about Shardy

* Quick fix for cuDecomp lowering

* Update src/jaxdecomp/_src/cudecomp/fft.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* add use_shardy argument to test_cudecomp_fft

* format

* Update tests/test_fft.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update src/jaxdecomp/_src/cudecomp/fft.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github
[…]
