+++ src/jaxdecomp/_src/spmd_ops.py
@@ -7,8 +7,12 @@ import jax.extend as jex
 from jax._src import custom_api_util, dispatch
 from jax.experimental.custom_partitioning import custom_partitioning
 from jax.interpreters import ad, batching, mlir, xla
+from packaging.version import Version as _Version
 
-# Imports
+if _Version(jax.__version__) >= _Version('0.7.0'):
+    ALLOW_SHARDY_PARTITIONER = True
+else:
+    ALLOW_SHARDY_PARTITIONER = False
 
 
 class BasePrimitive(metaclass=ABCMeta):
@@ -71,6 +75,33 @@ class BasePrimitive(metaclass=ABCMeta):
         return NotImplemented
 
     @staticmethod
+    def sharding_rule_producer(*args, **kwargs) -> Any:
+        """
+        Produces sharding rule for custom partitioning with Shardy partitioner.
+
+        This method should be implemented by subclasses to define how operations
+        should be partitioned across devices when using JAX's Shardy partitioner.
+
+        Parameters
+        ----------
+        *args
+            Variable positional arguments specific to the operation.
+        **kwargs
+            Variable keyword arguments specific to the operation.
+
+        Returns
+        -------
+        Any
+            Sharding rule specification (typically an einsum string).
+
+        Notes
+        -----
+        Base implementation returns NotImplemented and should be overridden
+        by concrete primitive implementations.
+        """
+        return NotImplemented
+
+    @staticmethod
     @abstractmethod
     def partition(*args, **kwargs) -> Any:
         """
@@ -120,6 +151,7 @@ def register_primitive(cls: type[BasePrimitive]) -> None:
         outer_p_lower.def_partition(
             infer_sharding_from_operands=cls.infer_sharding_from_operands,
             partition=cls.partition,
+            sharding_rule=cls.sharding_rule_producer,
         )
         mlir.register_lowering(
             outer_p,
@@ -154,28 +186,43 @@ class custom_spmd_rule:
         # Functions to be registered
         self.partition = None
         self.infer_sharding_from_operands = None
+        self.sharding_rule_producer = None
         self.jvp_rule = None
         self.transpose_rule = None
         self.batching_rule = None
 
     def def_partition(self, partition):
         self.partition = partition
-        if self.infer_sharding_from_operands is not None:
-            self.def_spmd_rule(partition, self.infer_sharding_from_operands)
+        if self.sharding_rule_producer is not None and jax.config.jax_use_shardy_partitioner:
+            self.def_spmd_rule(partition, self.infer_sharding_from_operands, self.sharding_rule_producer)
+        if self.infer_sharding_from_operands is not None and not jax.config.jax_use_shardy_partitioner:
+            self.def_spmd_rule(partition, self.infer_sharding_from_operands, self.sharding_rule_producer)
 
     def def_infer_sharding(self, infer_sharding_from_operands):
         self.infer_sharding_from_operands = infer_sharding_from_operands
         if self.partition is not None:
-            self.def_spmd_rule(self.partition, infer_sharding_from_operands)
+            self.def_spmd_rule(self.partition, infer_sharding_from_operands, self.sharding_rule_producer)
 
-    def def_spmd_rule(self, partition_rule, infer_sharding_rule):
+    def def_sharding_rule(self, sharding_rule_producer):
+        self.sharding_rule_producer = sharding_rule_producer
+        if self.partition is not None:
+            self.def_spmd_rule(self.partition, self.infer_sharding_from_operands, sharding_rule_producer)
+
+    def def_spmd_rule(self, partition_rule, infer_sharding_rule, sharding_rule_producer):
         assert partition_rule is not None, 'Partition rule is required'
[…]
