### 2025-09-22 b372183
Shardy Partionner must not be used prior to jax 0.7.0 including 0.7.0 (#90)
