### 2025-09-22 164ede3
Updating README.md with install instructions for cuDecomp backend
