### 2025-09-22 1df3b58
change demo.py to your script in README.md
