### 2025-09-22 e8dab75
Updating README.md with install instructions for cuDecomp backend (#91)
