### 2025-09-24 ff82a34
update pre-commit with mypy and update parameters names in tests for easier reading
