### 2025-09-24 a62390f
fix pyproject.toml typo
