+++ pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
 
 [project.optional-dependencies]
 test = [
-    'pytest>=8.0.0"',
+    'pytest>=8.0.0',
     'pytest-cov',
     'pytest-mock',
     'chex',
