### 2025-09-24 809c839
Change test paramters for better reading and add pytest-cov to test matrix (#92)
* Updating README.md with install instructions for cuDecomp backend

* change demo.py to your script in README.md

* update pre-commit with mypy and update parameters names in tests for easier reading

* fix pyproject.toml typo
