### 2025-09-30 DifferentiableUniverseInitiative/jaxDecomp#83 @ASKabalan
I am closing this issue
Feel free to open it if you still think that there is an issue
