### 2025-10-10 f30b58d
Add new CITATION.cff file (#93)
