### 2025-10-24 e7f32ff
fix errors and deprecation warning from jax 0.8.0
