+++ examples/lpt_nbody_demo.py
@@ -16,7 +16,7 @@ import jax_cosmo as jc
 import numpy as np
 from jax._src import mesh as mesh_lib
 from jax.experimental import mesh_utils
-from jax.experimental.shard_map import shard_map
+from jax import shard_map
 from jax.sharding import Mesh
 from jax.sharding import PartitionSpec as P
 from scatter import scatter
