+++ tests/test_halo.py
@@ -16,9 +16,8 @@ from math import prod
 
 import jax.numpy as jnp
 import pytest
-from jax import lax
+from jax import lax, shard_map
 from jax.experimental.multihost_utils import process_allgather
-from jax import shard_map
 from jax.sharding import PartitionSpec as P
 
 import jaxdecomp
