### 2025-10-24 eb2146d
Update jaxDecomp to fix errors and deprecations introduced in jax 0.8.0 (#94)
fix errors and deprecation warning from jax 0.8.0
