+++ CITATION.cff
@@ -30,5 +30,5 @@ keywords:
   - halo exchange
   - HPC
 license: MIT
-version: 0.2.8
+version: 0.2.9
 date-released: '2025-01-01'
