### 2025-11-13 #8852 @ASKabalan
> Hi @MoraruMaxim how is your review going?

He left me this comment
https://github.com/DifferentiableUniverseInitiative/jaxDecomp/issues/83
I will deal with it ASAP
