### 2025-11-27 DifferentiableUniverseInitiative/jaxDecomp#83 @ASKabalan
> This new graph doesn’t really solve the issue (it is still not weak scaling). In both plots, the workload per GPU isn’t constant, so they don’t actually represent weak scaling tests. I would suggest either re-running the tests with a constant workload per GPU or presenting the results differently, without labeling them as weak scaling.
>
> Also, I would really appreciate having the parallel efficiency plotted. Otherwise it’s quite hard to tell how good the scaling actually is.

I would have to relaunch the jobs

So to be sure I understand this correctly.
I need to have for example :

`N` in here is the number of elements of a 3D Grid with shape `X Y Z `
So `N` is `X*Y*Z`

I need to run :

| GPUs | Decomp | N        | X    | Y   | Z   |
|------|--------|----------|------|-----|-----|
| 2    | N/A    | 2097152  | 128  | 128 | 128 |
| 4    | 4x1    | 4194304  | 256  | 128 | 128 |
| 8    | 8x1    | 8388608  | 512  | 128 | 128 |
| 16   | 16x1   | 16777216 | 1024 | 128 | 128 |
| 32   | 32x1   | 33554432 | 2048 | 128 | 128 |
| 64   | 64x1   | 67108864 | 4096 | 128 | 128 |
| 128  | 128x1  | 134217728| 8192 | 128 | 128 |
| 256  | 256x1  | 268435456| 16384| 128 | 128 |

In this case the workload per GPU stays as `128x128x128`

Is this ok?
