### 2025-12-11 #8852 @ASKabalan
@mikaem Hello Mikael.
Would you kindly review this submission please.
I ask you since you are the author of [mpi4py-fft](https://github.com/mpi4py/mpi4py-fft) and my package has a similar purpose.

Thank you
