### 2026-01-22 #8852 @MoraruMaxim
At this stage, the main blocking point for my review is that I am unable to confirm the performance results reported in the paper. I was not able to successfully install the software with MPI or NCCL support (cuDecomp backend) and therefore could not run the provided benchmarks. I plan to retry the installation on a different machine and will report the encountered errors if the issue persists.
