### 2026-01-23 #8852 @ASKabalan
> At this stage, the main blocking point for my review is that I am unable to confirm the performance results reported in the paper. I was not able to successfully install the software with MPI or NCCL support (cuDecomp backend) and therefore could not run the provided benchmarks. I plan to retry the installation on a different machine and will report the encountered errors if the issue persists.

Also could you please confirm the table I made in this answer

https://github.com/DifferentiableUniverseInitiative/jaxDecomp/issues/83#issuecomment-3585063031

So I can launch the benchmarks on my end aswell

Thank you
