### 2026-01-29 #8852 @ASKabalan
@diehlpk
I will reproduce the perf tests very soon using the new table I suggested.
Since people are struggeling to install this with CUDA hardware .. can I suggest that people test the pure JAX backend (much easier to install)

This package also works on google colab TPU (and it is distributed accross TPU nodes aswell)
I can make a google colab so people can test the full functionalities without the hassle.
