### 2026-01-29 DifferentiableUniverseInitiative/jaxDecomp#83 @MoraruMaxim
This looks good for a weak scaling study! (constant workload per GPU: `64x128x128`)
