### 2026-01-29 #8852 @MoraruMaxim
> > At this stage, the main blocking point for my review is that I am unable to confirm the performance results reported in the paper. I was not able to successfully install the software with MPI or NCCL support (cuDecomp backend) and therefore could not run the provided benchmarks. I plan to retry the installation on a different machine and will report the encountered errors if the issue persists.
>
> Also could you please confirm the table I made in this answer
>
> [DifferentiableUniverseInitiative/jaxDecomp#83 (comment)](https://github.com/DifferentiableUniverseInitiative/jaxDecomp/issues/83#issuecomment-3585063031)
>
> So I can launch the benchmarks on my end aswell
>
> Thank you

My apologies for the delayed response. I was out of the office on vacation and am now catching up.

The table looks good to me!

> [@diehlpk](https://github.com/diehlpk) I will reproduce the perf tests very soon using the new table I suggested. Since people are struggeling to install this with CUDA hardware .. can I suggest that people test the pure JAX backend (much easier to install)
>
> This package also works on google colab TPU (and it is distributed accross TPU nodes aswell) I can make a google colab so people can test the full functionalities without the hassle.

@diehlpk I was able to test the pure JAX backend and verify most of the functionality; however, I was not able to independently reproduce the reported performance results. Does this pose any concern for the JOSS submission?
