### 2026-02-02 9b42c1b
refactor(pencil): improve pencil detection and add validation
