+++ src/jaxdecomp/_src/pencil_utils.py
@@ -62,7 +62,7 @@ def get_pdims_from_mesh(mesh: Optional[Mesh]) -> PdimsType:
     return pdims
 
 
-def get_pencil_type_from_mesh(mesh: Mesh) -> str:
+def get_pencil_type(mesh: Mesh) -> str:
     if mesh.empty:
         pdims = (1, 1)
     else:
@@ -94,7 +94,21 @@ def get_axis_names_from_mesh(mesh: Mesh) -> tuple[str, str]:
     return mesh.axis_names + (None,) * (2 - len(mesh.axis_names))
 
 
-# get_pdims_from_axis_names and get_pencil_type_from_axis_names are to be used UNDER SHARDMAP ONLY
+def get_axis_names_from_spec(mesh) -> tuple[AxisName, AxisName]:
+    """Returns the (x_axis, y_axis) names from a PartitionSpec.
+
+    This is the preferred way to determine axis names for 1D and 2D meshes,
+    as it correctly handles 1D meshes where the single mesh axis may map
+    to any array dimension.
+
+    Args:
+        spec: The PartitionSpec.
+
+    Returns:
+        A tuple of (x_axis_name, y_axis_name).
+    """
+    spec = spec + (None,) * (3 - len(spec))
+    return (spec[0], spec[1])
 
 
 def get_pdims_from_axis_names(x_axis_name: AxisName, y_axis_name: AxisName) -> PdimsType:
@@ -109,12 +123,44 @@ def get_pencil_type_from_axis_names(x_axis_name: AxisName, y_axis_name: AxisName
     return get_pencil_type_from_pdims(pdims)
 
 
-def get_transpose_order(fft_type: FftType, mesh: Optional[Mesh] = None) -> tuple[int, int, int]:
+def validate_spec_matches_mesh(spec, mesh: Mesh):
+    """Validate that the sharding spec is equivalent to the mesh's natural partitioning.
+
+    Equivalence rules (checked position by position):
+    - spec[i] == mesh.axis_names[i]  ->  direct match
+    - spec[i] is None and mesh.shape[mesh.axis_names[i]] == 1  ->  OK (size-1 axis = unsharded)
+    - spec[i] is None and i >= len(mesh.axis_names)  ->  OK (trailing unsharded dim)
+    - Otherwise  ->  ERROR (wrong order or wrong axis name)
+    """
+    spec = spec + (None,) * (3 - len(spec))
+    mesh_names = mesh.axis_names
+
+    for i in range(min(len(spec), 3)):
+        if spec[i] is not None:
+            if i >= len(mesh_names) or spec[i] != mesh_names[i]:
+                raise ValueError(
+                    f"Spec {spec} does not match mesh axis names {mesh_names}. "
+                    f"At position {i}: spec has '{spec[i]}' but mesh expects "
+                    f"'{mesh_names[i] if i < len(mesh_names) else 'nothing (trailing)'}'. "
+                    "Ensure the array is sharded with a spec matching the mesh axis order."
+                )
+        else:
+            if i < len(mesh_names) and mesh.shape[mesh_names[i]] > 1:
+                raise ValueError(
+                    f"Spec {spec} does not match mesh axis names {mesh_names}. "
+                    f"At position {i}: spec has None but mesh axis '{mesh_names[i]}' "
+                    f"has size {mesh.shape[mesh_names[i]]} (> 1). "
+                    "An unsharded dimension cannot correspond to a distributed mesh axis."
+                )
+
+
+def get_transpose_order(fft_type: FftType) -> tuple[int, int, int]:
     """Returns the transpose order based on the FFT type and mesh configuration.
 
     Args:
         fft_type: The type of FFT (FftType).
         mesh: The device mesh (Optional[Mesh]).
[…]
