### 2026-02-02 dda4658
feat(backend): propagate pencil_type and cleanup redundant logic
