### 2026-02-02 f7e4187
test(infra): refactor mesh and array creation for better flexibility
