+++ tests/conftest.py
@@ -44,7 +44,6 @@ def setup_and_teardown_session():
     yield
 
     import jax
-
     import jaxdecomp
 
     # Code to run at the end of the session
@@ -54,6 +53,11 @@ def setup_and_teardown_session():
 
     # Teardown code here
     # e.g., closing connections, cleaning up resources, etc.
+
+
+@pytest.fixture(params=['auto', 'explicit'])
+def axis_type(request):
+    return request.param
 
 
 def compare_sharding(sharding1, sharding2):
@@ -92,15 +96,35 @@ def process_slices(slices_tuple):
     return int(start_product + stop_product)
 
 
-def device_arange(pdims):
+def create_mesh(pdims, axis_type, transposed_devices=False):
+    import jax
+    from jax.experimental import mesh_utils
+    from jax.sharding import Mesh, AxisType
+
+
+    size = jax.device_count()
+    assert len(pdims) == 2
+    assert prod(pdims) == size, 'The product of pdims must be equal to the number of MPI processes'
+
+    # Remap to the global array from the local slicei
+    devices = mesh_utils.create_device_mesh(pdims)
+    if transposed_devices:
+        devices = devices.T
+
+    if axis_type == 'explicit':
+        axis_types = (AxisType.Explicit,) * len(pdims)
+    else:
+        axis_types = (AxisType.Auto,) * len(pdims)
+    return jax.make_mesh(devices.shape, ('z', 'y'), devices=devices.flatten(), axis_types=axis_types)
+
+
+
+def device_arange(mesh):
     import jax
     from jax import numpy as jnp
-    from jax.experimental import mesh_utils
-    from jax.sharding import Mesh, NamedSharding
+    from jax.sharding import NamedSharding
     from jax.sharding import PartitionSpec as P
 
-    devices = mesh_utils.create_device_mesh(pdims)
-    mesh = Mesh(devices.T, axis_names=('z', 'y'))
     sharding = NamedSharding(mesh, P('z', 'y'))
 
     def generate_aranged(x):
@@ -115,53 +139,39 @@ def device_arange(pdims):
     return aranged
 
 
-def create_ones_spmd_array(global_shape, pdims):
+def create_ones_spmd_array(global_shape, mesh):
     import jax
-    from jax.experimental import mesh_utils
-    from jax.sharding import Mesh, NamedSharding
+    from jax.sharding import NamedSharding
     from jax.sharding import PartitionSpec as P
 
-    size = jax.device_count()
     assert len(global_shape) == 3
-    assert len(pdims) == 2
-    assert prod(pdims) == size, 'The product of pdims must be equal to the number of MPI processes'
 
     local_shape = (
-        global_shape[0] // pdims[1],
-        global_shape[1] // pdims[0],
+        global_shape[0] // mesh.shape['z'],
+        global_shape[1] // mesh.shape['y'],
         global_shape[2],
     )
 
-    # Remap to the global array from the local slice
-    devices = mesh_utils.create_device_mesh(pdims)
-    mesh = Mesh(devices.T, axis_names=('z', 'y'))
     sharding = NamedSharding(mesh, P('z', 'y'))
     global_array = jax.make_array_from_callback(global_shape, sharding, data_callback=lambda _: jax.numpy.ones(local_shape))
 
-    return global_array, mesh
+    return global_array
 
 
 # Helper function to create a 3D array and remap it to the global array
-def create_spmd_array(global_shape, pdims):
+def create_spmd_array(global_shape, mesh):
     import jax
-    from jax.experimental import mesh_utils
-    from jax.sharding import Mesh, NamedSharding
+    from jax.sharding import NamedSharding
[…]
