### 2026-02-02 a32c5c1
test: update tests to support axis_type parameterization
