### 2026-02-02 #8852 @diehlpk
@MoraruMaxim

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

One check box is about performance, see above.

@ASKabalan if reviewers can not reproduce the performance, we might need to remove them from the paper.
