### 2026-02-02 #8852 @ASKabalan
I can produce new plots this week.
However I will run them on 256 GPUs (which is what this package is made for )

One can in theory check the performance on Google colab TPUs ... But this is not what this package is all about.

If needed I can remove them from the paper but keep them in the readme so users know that this package is meant for scalability
