### 2026-02-02 #8852 @diehlpk
@ASKabalan You can go as far as you want. It is ok if the reviewers can reproduce a small subset for your data. Even from one to 16 GPUs would be ok.
