### 2026-02-02 #8852 @ASKabalan
> [@ASKabalan](https://github.com/ASKabalan) You can go as far as you want. It is ok if the reviewers can reproduce a small subset for your data. Even from one to 16 GPUs would be ok.

Ok.

I am updating the package for Jax v0.9.0 this week
I will update the installation process and make a script that reviewers can use.
