### 2026-02-03 #8852 @berkhess
> Hi [@berkhess](https://github.com/berkhess) do you have time to review this paper?

I might. It is unclear what is expected from me though. I do not have time to compile and test the software.

I had a quick a look. The 3D-FFT using cuFFT is a factor 10 slower at scale than Nvidia's own numbers on Selene, also using A100. This might be because of lack of NV-link, but then that should be checked and made clear.
