### 2026-02-04 309d1f7
refactor: improve pencil detection and add sharding spec validation
