+++ src/jaxdecomp/__init__.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from importlib.metadata import PackageNotFoundError, version
 
-from jaxdecomp._src.pencil_utils import get_fft_output_sharding, get_output_specs
+from jaxdecomp._src.pencil_utils import get_fft_output_sharding, get_output_specs, validate_spec_matches_mesh
 from jaxdecomp.fft import fftfreq3d, pfft3d, pifft3d, rfftfreq3d
 from jaxdecomp.halo import halo_exchange
 from jaxdecomp.transpose import (
@@ -75,6 +75,7 @@ __all__ = [
     'NO_DECOMP',
     'get_fft_output_sharding',
     'get_output_specs',
+    'validate_spec_matches_mesh',
     'fftfreq3d',
     'rfftfreq3d',
     'HALO_COMM_MPI',
