### 2026-02-04 29583b6
refactor: update backends to use new pencil utilities
