### 2026-02-04 b79843c
test: update tests for new API and validation logic
