### 2026-02-04 43fbe3e
docs: update installation guide and configuration
