### 2026-02-04 e115cc0
chore: revamp benchmark scripts
