+++ benchmarks/jaxfft.py
@@ -1,157 +1,0 @@
-import jax
-
-jax.distributed.initialize()
-rank = jax.process_index()
-size = jax.process_count()
-import argparse
-
-import jax.numpy as jnp
-from jax.experimental import mesh_utils, multihost_utils
-from jax.experimental.multihost_utils import sync_global_devices
-from jax.sharding import Mesh
-from jax.sharding import PartitionSpec as P
-from jax_hpc_profiler import Timer
-
-import jaxdecomp
-
-
-def run_benchmark(pdims, global_shape, backend, nb_nodes, precision, iterations, trace, contiguous, output_path):
-    pfft_backend = 'cudecomp'
-    if backend == 'NCCL':
-        jaxdecomp.config.update('transpose_comm_backend', jaxdecomp.TRANSPOSE_COMM_NCCL)
-    elif backend == 'NCCL_PL':
-        jaxdecomp.config.update('transpose_comm_backend', jaxdecomp.TRANSPOSE_COMM_NCCL_PL)
-    elif backend == 'MPI_P2P':
-        jaxdecomp.config.update('transpose_comm_backend', jaxdecomp.TRANSPOSE_COMM_MPI_P2P)
-    elif backend == 'MPI':
-        jaxdecomp.config.update('transpose_comm_backend', jaxdecomp.TRANSPOSE_COMM_MPI_A2A)
-    else:
-        pfft_backend = 'jax'
-    # Initialize the local slice with the local slice shape
-    array = jax.random.normal(shape=[global_shape[0] // pdims[0], global_shape[1] // pdims[1], global_shape[2]], key=jax.random.PRNGKey(rank))
-
-    if contiguous:
-        jaxdecomp.config.update('transpose_axis_contiguous', contiguous)
-
-    # Remap to the global array from the local slice
-    devices = mesh_utils.create_device_mesh(pdims)
-    mesh = Mesh(devices.T, axis_names=('z', 'y'))
-    global_array = multihost_utils.host_local_array_to_global_array(array, mesh, P('z', 'y'))
-
-    if jax.process_index() == 0:
-        print(f'Devices {jax.devices()}')
-        print(
-            f'Global dims {global_shape}, pdims {pdims} , Backend {backend} local array shape {array.shape} global array shape {global_array.shape}'
-        )
-        print('Sharding :')
-        print(global_array.sharding)
-
-    @jax.jit
-    def do_fft(x):
-        return jaxdecomp.fft.pfft3d(x, backend=pfft_backend)
-
-    @jax.jit
-    def do_ifft(x):
-        return jaxdecomp.fft.pifft3d(x, backend=pfft_backend)
-
-    @jax.jit
-    def get_diff(arr1, arr2):
-        return jnp.abs(arr1 - arr2).max()
-
-    fft_chrono = Timer(save_jaxpr=False)
-    ifft_chrono = Timer(save_jaxpr=False)
-
-    # Warm start
-    global_array = fft_chrono.chrono_jit(do_fft, global_array)
-    global_array = ifft_chrono.chrono_jit(do_ifft, global_array)
-    sync_global_devices('warmup')
-    for i in range(iterations):
-        global_array = fft_chrono.chrono_fun(do_fft, global_array)
-        global_array = ifft_chrono.chrono_fun(do_ifft, global_array)
-
-    cont_str = '-cont' if contiguous else '-noncont'
-    fft_metadata = {
-        'function': f'FFT{cont_str}',
-        'precision': precision,
-        'x': str(global_shape[0]),
-        'px': str(pdims[0]),
-        'py': str(pdims[1]),
-        'backend': backend,
-        'nodes': str(nb_nodes),
-    }
-    ifft_metadata = {
[…]
