### 2026-02-04 c5e9503
remove set_mesh from test
