### 2026-02-04 c8bf5eb
Skip explicit axis type
