### 2026-02-04 b472f12
Update for joss review 2 (#98)
* refactor: improve pencil detection and add sharding spec validation

* refactor: update backends to use new pencil utilities

* test: update tests for new API and validation logic

* docs: update installation guide and configuration

* chore: revamp benchmark scripts

* Format

* remove set_mesh from test

* Skip explicit axis type
