### 2026-02-04 306c9a8
refactor: improve validation and error handling in cuDecomp backend
