### 2026-02-04 fa784ff
refactor: improve validation and error handling in cuDecomp backend (#102)
