### 2026-02-04 662fe3f
chore: update cuDecomp submodule and adapt C++ calls to new API
