+++ src/csrc/halo.cu
@@ -21,8 +21,8 @@ HRESULT HaloExchange<real_t>::get_halo_descriptor(cudecompHandle_t handle, size_
   CHECK_CUDECOMP_EXIT(cudecompGridDescCreate(handle, &m_GridConfig, &config, nullptr));
 
   // Get pencil information for the specified axis
-  CHECK_CUDECOMP_EXIT(
-      cudecompGetPencilInfo(handle, m_GridConfig, &m_PencilInfo, halo_desc.axis, halo_desc.halo_extents.data()));
+  CHECK_CUDECOMP_EXIT(cudecompGetPencilInfo(handle, m_GridConfig, &m_PencilInfo, halo_desc.axis,
+                                            halo_desc.halo_extents.data(), nullptr));
 
   // Get workspace size
   int64_t workspace_num_elements;
@@ -53,15 +53,18 @@ HRESULT HaloExchange<real_t>::halo_exchange(cudecompHandle_t handle, haloDescrip
     switch (desc.axis) {
     case 0:
       CHECK_CUDECOMP_EXIT(cudecompUpdateHalosX(handle, m_GridConfig, data_d, work_d, get_cudecomp_datatype(real_t(0)),
-                                               m_PencilInfo.halo_extents, desc.halo_periods.data(), i, stream));
+                                               m_PencilInfo.halo_extents, desc.halo_periods.data(), i, nullptr,
+                                               stream));
       break;
     case 1:
       CHECK_CUDECOMP_EXIT(cudecompUpdateHalosY(handle, m_GridConfig, data_d, work_d, get_cudecomp_datatype(real_t(0)),
-                                               m_PencilInfo.halo_extents, desc.halo_periods.data(), i, stream));
+                                               m_PencilInfo.halo_extents, desc.halo_periods.data(), i, nullptr,
+                                               stream));
       break;
     case 2:
       CHECK_CUDECOMP_EXIT(cudecompUpdateHalosZ(handle, m_GridConfig, data_d, work_d, get_cudecomp_datatype(real_t(0)),
-                                               m_PencilInfo.halo_extents, desc.halo_periods.data(), i, stream));
+                                               m_PencilInfo.halo_extents, desc.halo_periods.data(), i, nullptr,
+                                               stream));
       break;
     }
   }
