### 2026-02-04 bf0ce24
Update cudecomp cpp (#103)
* build: update RPATH settings for NVHPC math libraries

* chore: update cuDecomp submodule and adapt C++ calls to new API
