+++ src/csrc/fft.cu
@@ -20,11 +20,11 @@ HRESULT FourierExecutor<real_t>::Initialize(cudecompHandle_t handle, size_t& wor
 
   CHECK_CUDECOMP_EXIT(cudecompGridDescCreate(handle, &m_GridConfig, &m_GridDescConfig, nullptr));
   // Get x-pencil information (complex)
-  CHECK_CUDECOMP_EXIT(cudecompGetPencilInfo(handle, m_GridConfig, &m_XPencilInfo, 0, nullptr));
+  CHECK_CUDECOMP_EXIT(cudecompGetPencilInfo(handle, m_GridConfig, &m_XPencilInfo, 0, nullptr, nullptr));
   // Get y-pencil information (complex)
-  CHECK_CUDECOMP_EXIT(cudecompGetPencilInfo(handle, m_GridConfig, &m_YPencilInfo, 1, nullptr));
+  CHECK_CUDECOMP_EXIT(cudecompGetPencilInfo(handle, m_GridConfig, &m_YPencilInfo, 1, nullptr, nullptr));
   // Get z-pencil information (complex)
-  CHECK_CUDECOMP_EXIT(cudecompGetPencilInfo(handle, m_GridConfig, &m_ZPencilInfo, 2, nullptr));
+  CHECK_CUDECOMP_EXIT(cudecompGetPencilInfo(handle, m_GridConfig, &m_ZPencilInfo, 2, nullptr, nullptr));
   // Get workspace size
   int64_t num_elements_work_c;
   CHECK_CUDECOMP_EXIT(cudecompGetTransposeWorkspaceSize(handle, m_GridConfig, &num_elements_work_c));
@@ -303,7 +303,8 @@ HRESULT FourierExecutor<real_t>::forwardXY(cudecompHandle_t handle, fftDescripto
     CHECK_CUFFT_EXIT(cufftXtExec(m_Plan_c2c_yz, input, output, DIRECTION));
     // Tranpose Y to X but it actually X to Z
     CHECK_CUDECOMP_EXIT(cudecompTransposeYToX(handle, m_GridConfig, output, output, work_d,
-                                              get_cudecomp_datatype(complex_t(0)), nullptr, nullptr, stream));
+                                              get_cudecomp_datatype(complex_t(0)), nullptr, nullptr, nullptr, nullptr,
+                                              stream));
     CHECK_CUFFT_EXIT(cufftSetStream(m_Plan_c2c_x, stream));
     CHECK_CUFFT_EXIT(cufftSetWorkArea(m_Plan_c2c_x, work_d));
     // FFT on the second slab
@@ -318,10 +319,12 @@ HRESULT FourierExecutor<real_t>::forwardXY(cudecompHandle_t handle, fftDescripto
     CHECK_CUFFT_EXIT(cufftXtExec(m_Plan_c2c_xy, input, output, DIRECTION));
     // Tranpose X to Y
     CHECK_CUDECOMP_EXIT(cudecompTransposeXToY(handle, m_GridConfig, output, output, work_d,
-                                              get_cudecomp_datatype(complex_t(0)), nullptr, nullptr, stream));
+                                              get_cudecomp_datatype(complex_t(0)), nullptr, nullptr, nullptr, nullptr,
+                                              stream));
     // Tranpose Y to z
     CHECK_CUDECOMP_EXIT(cudecompTransposeYToZ(handle, m_GridConfig, output, output, work_d,
-                                              get_cudecomp_datatype(complex_t(0)), nullptr, nullptr, stream));
+                                              get_cudecomp_datatype(complex_t(0)), nullptr, nullptr, nullptr, nullptr,
+                                              stream));
[…]
