### 2026-02-04 3bfc10b
update docs and README to use AxisType.Auto in mesh creation
