### 2026-02-04 e01df85
update docs and README to use AxisType.Auto in mesh creation (#104)
