### 2026-02-06 f69bc53
docs: add report for JAX custom partitioning bug with explicit axes
