### 2026-02-06 ab6c489
refactor: migrate backend to JAX FFI and nanobind
