+++ src/jaxdecomp/_src/cudecomp/halo.py
@@ -2,11 +2,10 @@ from functools import partial
 from typing import Any
 
 import jax
-import jaxlib.mlir.ir as ir
+import jax.ffi
 import numpy as np
 from jax import ShapeDtypeStruct
 from jax._src.interpreters import mlir
-from jax._src.interpreters.mlir import custom_call
 from jax._src.typing import Array
 from jax.core import ShapedArray
 from jax.sharding import Mesh, NamedSharding
@@ -41,7 +40,7 @@ class HaloPrimitive(BasePrimitive):
         halo_periods: Periodicity,
         pdims: PdimsType,
         global_shape: GdimsType,
-    ) -> ShapedArray:
+    ) -> list[ShapedArray]:
         """
         Abstract function for determining the shape and dtype after the halo exchange operation.
 
@@ -60,11 +59,28 @@ class HaloPrimitive(BasePrimitive):
 
         Returns
         -------
-        Array
-            Abstract array after the halo exchange operation.
+        list[ShapedArray]
+            Workspace aval and output aval.
         """
-        del halo_extents, halo_periods, pdims, global_shape
-        return ShapedArray(x.shape, x.dtype)
+        output_aval = ShapedArray(x.shape, x.dtype)
+
+        is_double = np.finfo(x.dtype).dtype == np.float64
+        lowered_halo_extents = (*halo_extents, 0)
+        lowered_halo_periods = (*halo_periods, True)
+
+        workspace_size = _jaxdecomp.get_halo_workspace_size(
+            gdims=list(global_shape[::-1]),
+            pdims=list(pdims),
+            transpose_comm_backend=int(jaxdecomp.config.transpose_comm_backend),
+            halo_comm_backend=int(jaxdecomp.config.halo_comm_backend),
+            halo_extents=list(lowered_halo_extents[::-1]),
+            halo_periods=list(lowered_halo_periods[::-1]),
+            axis=0,
+            double_precision=is_double,
+        )
+        workspace_aval = ShapedArray([workspace_size], np.uint8)
+
+        return [workspace_aval, output_aval]
 
     @staticmethod
     def outer_abstract(x: Array, halo_extents: HaloExtentType, halo_periods: Periodicity) -> ShapedArray:
@@ -91,12 +107,12 @@ class HaloPrimitive(BasePrimitive):
     @staticmethod
     def lowering(
         ctx,
-        x: ir.Value,
+        x,
         halo_extents: HaloExtentType,
         halo_periods: Periodicity,
         pdims: PdimsType,
         global_shape: GdimsType,
-    ) -> ir.OpResultList:
+    ):
         """
         Lowering function to generate the MLIR representation for halo exchange.
 
@@ -104,7 +120,7 @@ class HaloPrimitive(BasePrimitive):
         ----------
         ctx : Any
             Context for the operation.
-        x : Array
+        x
             Input array.
         halo_extents : HaloExtentType
             Extents of the halo in x, y, and z dimensions.
@@ -117,41 +133,32 @@ class HaloPrimitive(BasePrimitive):
 
         Returns
         -------
-        Array
-            Resulting array after the halo exchange operation.
+        list
+            Resulting MLIR values after the halo exchange operation.
         """
         (x_aval,) = ctx.avals_in
-        x_type = ir.RankedTensorType(x.type)
-        n = len(x_type.shape)
-
         is_double = np.finfo(x_aval.dtype).dtype == np.float64
+        ffi_name = 'halo_C128' if is_double else 'halo_C64'
 
-        # Compute the descriptor for the halo exchange operation
-        config = _jaxdecomp.GridConfig()
-        config.pdims = pdims
[…]
