+++ src/jaxdecomp/_src/cudecomp/fft.py
@@ -2,12 +2,10 @@ from functools import partial
 from typing import Any
 
 import jax
-import jaxlib.mlir.ir as ir
+import jax.ffi
 import numpy as np
 from jax import ShapeDtypeStruct
 from jax._src.interpreters import mlir
-from jax._src.interpreters.mlir import custom_call
-from jax._src.lib.mlir.dialects import hlo
 from jax._src.numpy.util import promote_dtypes_complex
 from jax.core import ShapedArray
 from jax.extend.core import Primitive
@@ -53,7 +51,7 @@ class FFTPrimitive(BasePrimitive):
         global_shape: GdimsType,
         adjoint: bool,
         mesh: Mesh,
-    ) -> ShapedArray:
+    ) -> list[ShapedArray]:
         """
         Abstract function to compute the shape of FFT output.
 
@@ -74,22 +72,40 @@ class FFTPrimitive(BasePrimitive):
 
         Returns
         -------
-        ShapedArray
-            Shape of the output array.
+        list[ShapedArray]
+            Workspace aval and output aval.
         """
-        del mesh
         if global_shape == x.shape:
-            return FFTPrimitive.outer_abstract(x, fft_type=fft_type, adjoint=adjoint)
+            output_aval = FFTPrimitive.outer_abstract(x, fft_type=fft_type, adjoint=adjoint)
+        else:
+            transpose_shape = get_transpose_order(fft_type)
+            output_shape = (
+                global_shape[transpose_shape[0]] // pdims[0],
+                global_shape[transpose_shape[1]] // pdims[1],
+                global_shape[transpose_shape[2]] // pdims[2],
+            )
+            output_aval = ShapedArray(output_shape, x.dtype)
 
-        transpose_shape = get_transpose_order(fft_type)
+        is_double = np.finfo(x.dtype).dtype == np.float64
+        pencil_type = get_pencil_type(mesh)
+        forward = fft_type in (FftType.FFT,)
+        local_transpose = jaxdecomp.config.transpose_axis_contiguous
+        original_pdims, lowering_global_shape = get_lowering_args(fft_type, global_shape, mesh)
 
-        output_shape = (
-            global_shape[transpose_shape[0]] // pdims[0],
-            global_shape[transpose_shape[1]] // pdims[1],
-            global_shape[transpose_shape[2]] // pdims[2],
+        workspace_size = _jaxdecomp.get_fft_workspace_size(
+            gdims=list(lowering_global_shape[::-1]),
+            pdims=list(original_pdims),
+            transpose_comm_backend=int(jaxdecomp.config.transpose_comm_backend),
+            halo_comm_backend=int(jaxdecomp.config.halo_comm_backend),
+            forward=forward,
+            double_precision=is_double,
+            adjoint=adjoint,
+            contiguous=local_transpose,
+            decomposition=int(pencil_type),
         )
+        workspace_aval = ShapedArray([workspace_size], np.uint8)
 
-        return ShapedArray(output_shape, x.dtype)
+        return [workspace_aval, output_aval]
 
     @staticmethod
     def outer_abstract(x: Array, fft_type: FftType, adjoint: bool) -> ShapedArray:
@@ -127,14 +143,14 @@ class FFTPrimitive(BasePrimitive):
     @staticmethod
     def lowering(
         ctx,
-        a: ir.Value,
+        a,
         *,
         fft_type: FftType,
         pdims: TransposablePdimsType,
         global_shape: GdimsType,
         adjoint: bool,
         mesh: Mesh,
-    ) -> ir.OpResultList:
+    ):
         """
         Lowering function for FFT primitive.
 
[…]
