+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["scikit-build-core>=0.4.0", "pybind11>=2.9.0"]
+requires = ["scikit-build-core>=0.4.3", "nanobind>=1.3.2", "jax>=0.4.35"]
 build-backend = "scikit_build_core.build"
 [project]
 name = "jaxdecomp"

[…]
