### 2026-02-06 dda7bdf
Roll back workspace buffer to be input instead of output
