+++ src/jaxdecomp/_src/cudecomp/fft.py
@@ -3,9 +3,11 @@ from typing import Any
 
 import jax
 import jax.ffi
+import jaxlib.mlir.ir as ir
 import numpy as np
 from jax import ShapeDtypeStruct
 from jax._src.interpreters import mlir
+from jax._src.lib.mlir.dialects import hlo
 from jax._src.numpy.util import promote_dtypes_complex
 from jax.core import ShapedArray
 from jax.extend.core import Primitive
@@ -51,7 +53,7 @@ class FFTPrimitive(BasePrimitive):
         global_shape: GdimsType,
         adjoint: bool,
         mesh: Mesh,
-    ) -> list[ShapedArray]:
+    ) -> ShapedArray:
         """
         Abstract function to compute the shape of FFT output.
 
@@ -72,40 +74,19 @@ class FFTPrimitive(BasePrimitive):
 
         Returns
         -------
-        list[ShapedArray]
-            Workspace aval and output aval.
+        ShapedArray
+            Output aval.
         """
+        del mesh
         if global_shape == x.shape:
-            output_aval = FFTPrimitive.outer_abstract(x, fft_type=fft_type, adjoint=adjoint)
-        else:
-            transpose_shape = get_transpose_order(fft_type)
-            output_shape = (
-                global_shape[transpose_shape[0]] // pdims[0],
-                global_shape[transpose_shape[1]] // pdims[1],
-                global_shape[transpose_shape[2]] // pdims[2],
-            )
-            output_aval = ShapedArray(output_shape, x.dtype)
-
-        is_double = np.finfo(x.dtype).dtype == np.float64
-        pencil_type = get_pencil_type(mesh)
-        forward = fft_type in (FftType.FFT,)
-        local_transpose = jaxdecomp.config.transpose_axis_contiguous
-        original_pdims, lowering_global_shape = get_lowering_args(fft_type, global_shape, mesh)
-
-        workspace_size = _jaxdecomp.get_fft_workspace_size(
-            gdims=list(lowering_global_shape[::-1]),
-            pdims=list(original_pdims),
-            transpose_comm_backend=int(jaxdecomp.config.transpose_comm_backend),
-            halo_comm_backend=int(jaxdecomp.config.halo_comm_backend),
-            forward=forward,
-            double_precision=is_double,
-            adjoint=adjoint,
-            contiguous=local_transpose,
-            decomposition=int(pencil_type),
+            return FFTPrimitive.outer_abstract(x, fft_type=fft_type, adjoint=adjoint)
+        transpose_shape = get_transpose_order(fft_type)
+        output_shape = (
+            global_shape[transpose_shape[0]] // pdims[0],
+            global_shape[transpose_shape[1]] // pdims[1],
+            global_shape[transpose_shape[2]] // pdims[2],
         )
-        workspace_aval = ShapedArray([workspace_size], np.uint8)
-
-        return [workspace_aval, output_aval]
+        return ShapedArray(output_shape, x.dtype)
 
     @staticmethod
     def outer_abstract(x: Array, fft_type: FftType, adjoint: bool) -> ShapedArray:
@@ -178,7 +159,9 @@ class FFTPrimitive(BasePrimitive):
         """
         del pdims
         (x_aval,) = ctx.avals_in
+        (aval_out,) = ctx.avals_out
         dtype = x_aval.dtype
+        a_type = ir.RankedTensorType(a.type)
 
         assert fft_type in (
             FftType.FFT,
@@ -193,23 +176,54 @@ class FFTPrimitive(BasePrimitive):
         original_pdims, global_shape = get_lowering_args(fft_type, global_shape, mesh)
[…]
