### 2026-02-06 65bea80
Update tests and code for migrating to FFI
