+++ tests/test_fft.py
@@ -18,6 +18,8 @@ from functools import partial
 import jax.numpy as jnp
 import pytest
 from jax.experimental.multihost_utils import process_allgather
+from jax.sharding import NamedSharding, auto_axes
+from jax.sharding import PartitionSpec as P
 
 import jaxdecomp
 from jaxdecomp._src import PENCILS, SLAB_XY, SLAB_YZ
@@ -54,14 +56,11 @@ class TestFFTs:
         if use_shardy and not ALLOW_SHARDY_PARTITIONER:
             pytest.skip(reason='Shardy partitioner is not supported in this JAX version use at least JAX 0.7.0')
 
-        if axis_type == 'explicit':
-            pytest.skip(reason='Explicit axis type not yet supported for FFT tests')
-
         print('*' * 80)
         print(f'Testing with pdims {pdims} and global shape {global_shape} and local transpose {local_transpose} use shardy {use_shardy}')
         if pdims[0] == 1:
             penciltype = SLAB_XY
-        elif pdims[1] == 1:
+        elif len(pdims) < 2 or pdims[1] == 1:
             penciltype = SLAB_YZ
         else:
             penciltype = PENCILS
@@ -73,10 +72,29 @@ class TestFFTs:
         mesh = create_mesh(pdims, axis_type=axis_type)
         global_array = create_spmd_array(global_shape, mesh)
 
+        if axis_type == 'explicit':
+            # Perform distributed FFT with wrapper
+            out_sharding = jaxdecomp.get_fft_output_sharding(global_array.sharding)
+
+            @auto_axes
+            def pfft3d(x, out_sharding=out_sharding):
+                return jaxdecomp.fft.pfft3d(x, backend=backend)
+
+            # Perform inverse FFT with wrapper
+            @auto_axes
+            def pifft3d(x, out_sharding=global_array.sharding):
+                return jaxdecomp.fft.pifft3d(x, backend=backend)
+
+            pfft3d = partial(pfft3d, out_sharding=out_sharding)
+            pifft3d = partial(pifft3d, out_sharding=global_array.sharding)
+        else:
+            pfft3d = partial(jaxdecomp.fft.pfft3d, backend=backend)
+            pifft3d = partial(jaxdecomp.fft.pifft3d, backend=backend)
+
         # Perform distributed FFT
-        karray = jaxdecomp.fft.pfft3d(global_array, backend=backend)
+        karray = pfft3d(global_array)
         # Perform inverse FFT
-        rec_array = jaxdecomp.fft.pifft3d(karray, backend=backend)
+        rec_array = pifft3d(karray)
 
         print(f'orignal shard {global_array.sharding.spec}')
         print(f'sharding of karray {karray.sharding.spec}')
@@ -303,7 +321,7 @@ class TestFFTFreq:
     @pytest.mark.parametrize('pdims', decomp)  # Test with Slab and Pencil decompositions
     @pytest.mark.parametrize('global_shape', global_shapes)  # Test cubes, non-cubes and primes
     def test_cudecomp_fft(self, pdims, global_shape, local_transpose, use_shardy, axis_type):
-        self.run_test(pdims, global_shape, local_transpose, backend='cuDecomp', use_shardy=use_shardy, axis_type=axis_type)
+        self.run_test(pdims, global_shape, local_transpose, backend='cuDeComp', use_shardy=use_shardy, axis_type=axis_type)
 
     # Cartesian product tests
     @pytest.mark.parametrize('local_transpose', local_transpose)  # Test with and without local transpose
[…]
