+++ tests/test_transpose.py
@@ -18,8 +18,13 @@ from functools import partial
 import jax.numpy as jnp
 import pytest
 from jax.experimental.multihost_utils import process_allgather
-from jax.sharding import NamedSharding, auto_axes
+from jax.sharding import NamedSharding
 from jax.sharding import PartitionSpec as P
+
+try:
+    from jax.sharding import auto_axes
+except ImportError:
+    auto_axes = None
 
 import jaxdecomp
 from jaxdecomp import (
@@ -77,6 +82,8 @@ class TestTransposes:
             z_pencil_sharding = NamedSharding(mesh, P(None, 'z', 'y'))
 
         if axis_type == 'explicit':
+            if auto_axes is None:
+                pytest.skip(reason='auto_axes is not available in this JAX version, please upgrade to at least JAX 0.9.0')
 
             @auto_axes
             def transposeXtoY_safe(x, out_sharding=y_pencil_sharding):
