### 2026-02-06 9778004
Migrating to ffi (#105)
* docs: add report for JAX custom partitioning bug with explicit axes

* refactor: migrate backend to JAX FFI and nanobind

* Roll back workspace buffer to be input instead of output

* Update tests and code for migrating to FFI

* Delete file

* Deactivate explicit axes test on github

* Fix errors
