+++ src/jaxdecomp/_src/cudecomp/fft.py
@@ -2,12 +2,11 @@ from functools import partial
 from typing import Any
 
 import jax
+import jax.ffi
 import jaxlib.mlir.ir as ir
 import numpy as np
 from jax import ShapeDtypeStruct
 from jax._src.interpreters import mlir
-from jax._src.interpreters.mlir import custom_call
-from jax._src.lib.mlir.dialects import hlo
 from jax._src.numpy.util import promote_dtypes_complex
 from jax.core import ShapedArray
 from jax.extend.core import Primitive
@@ -75,20 +74,16 @@ class FFTPrimitive(BasePrimitive):
         Returns
         -------
         ShapedArray
-            Shape of the output array.
+            Output aval.
         """
         del mesh
-        if global_shape == x.shape:
-            return FFTPrimitive.outer_abstract(x, fft_type=fft_type, adjoint=adjoint)
 
         transpose_shape = get_transpose_order(fft_type)
-
         output_shape = (
             global_shape[transpose_shape[0]] // pdims[0],
             global_shape[transpose_shape[1]] // pdims[1],
             global_shape[transpose_shape[2]] // pdims[2],
         )
-
         return ShapedArray(output_shape, x.dtype)
 
     @staticmethod
@@ -127,14 +122,14 @@ class FFTPrimitive(BasePrimitive):
     @staticmethod
     def lowering(
         ctx,
-        a: ir.Value,
+        a,
         *,
         fft_type: FftType,
         pdims: TransposablePdimsType,
         global_shape: GdimsType,
         adjoint: bool,
         mesh: Mesh,
-    ) -> ir.OpResultList:
+    ):
         """
         Lowering function for FFT primitive.
 
@@ -142,7 +137,7 @@ class FFTPrimitive(BasePrimitive):
         ----------
         ctx
             Context.
-        a : Array
+        a
             Input array.
         fft_type : FftType
             Type of FFT operation (forward or inverse).
@@ -173,35 +168,49 @@ class FFTPrimitive(BasePrimitive):
 
         forward = fft_type in (FftType.FFT,)
         is_double = np.finfo(dtype).dtype == np.float64
+        ffi_name = 'pfft_C128' if is_double else 'pfft_C64'
 
         pencil_type = get_pencil_type(mesh)
         original_pdims, global_shape = get_lowering_args(fft_type, global_shape, mesh)
         local_transpose = jaxdecomp.config.transpose_axis_contiguous
 
-        config = _jaxdecomp.GridConfig()
-        config.pdims = original_pdims
-        config.gdims = global_shape[::-1]
-        config.halo_comm_backend = jaxdecomp.config.halo_comm_backend
-        config.transpose_comm_backend = jaxdecomp.config.transpose_comm_backend
-        workspace_size, opaque = _jaxdecomp.build_fft_descriptor(config, forward, is_double, adjoint, local_transpose, pencil_type)
+        workspace_size = _jaxdecomp.get_fft_workspace_size(
+            gdims=list(global_shape[::-1]),
+            pdims=list(original_pdims),
+            transpose_comm_backend=int(jaxdecomp.config.transpose_comm_backend.value),
+            halo_comm_backend=int(jaxdecomp.config.halo_comm_backend.value),
+            forward=forward,
+            double_precision=is_double,
+            adjoint=adjoint,
+            contiguous=local_transpose,
+            decomposition=int(pencil_type.value),
+        )
 
         n = len(a_type.shape)
         layout = tuple(range(n - 1, -1, -1))
-
         workspace = mlir.full_like_aval(ctx, 0, ShapedArray(shape=[workspace_size], dtype=np.byte))
 
-        result = custom_call(
-            'pfft3d',
-            result_types=[a_type],
-            operands=[a, workspace],
[…]
