### 2026-02-06 dc85729
Fix bug with transpose and halo exchange lowering
