### 2026-02-07 deeb41a
Fix bug with transpose and halo exchange lowering (#107)
