+++ src/csrc/jaxdecomp.cc
@@ -138,7 +138,7 @@ template <ffi::DataType T>
 ffi::Error pfft3d_ffi(cudaStream_t stream, ffi::Span<const int64_t> gdims, ffi::Span<const int64_t> pdims,
                       int64_t transpose_comm_backend, int64_t halo_comm_backend, bool forward, bool adjoint,
                       bool contiguous, int64_t decomposition, ffi::Buffer<T> input,
-                      ffi::Buffer<ffi::DataType::U8> workspace, ffi::Result<ffi::Buffer<T>> output) {
+                      ffi::Buffer<ffi::DataType::S8> workspace, ffi::Result<ffi::Buffer<T>> output) {
   if (gdims.size() != 3) { return ffi::Error::InvalidArgument("gdims must have 3 elements"); }
   if (pdims.size() != 2) { return ffi::Error::InvalidArgument("pdims must have 2 elements"); }
 
@@ -185,7 +185,7 @@ template <ffi::DataType T>
 ffi::Error halo_ffi(cudaStream_t stream, ffi::Span<const int64_t> gdims, ffi::Span<const int64_t> pdims,
                     int64_t transpose_comm_backend, int64_t halo_comm_backend, ffi::Span<const int64_t> halo_extents,
                     ffi::Span<const int64_t> halo_periods, int64_t axis, ffi::Buffer<T> input,
-                    ffi::Buffer<ffi::DataType::U8> workspace, ffi::Result<ffi::Buffer<T>> output) {
+                    ffi::Buffer<ffi::DataType::S8> workspace, ffi::Result<ffi::Buffer<T>> output) {
   if (gdims.size() != 3) { return ffi::Error::InvalidArgument("gdims must have 3 elements"); }
   if (pdims.size() != 2) { return ffi::Error::InvalidArgument("pdims must have 2 elements"); }
   if (halo_extents.size() != 3) { return ffi::Error::InvalidArgument("halo_extents must have 3 elements"); }
@@ -235,7 +235,7 @@ ffi::Error halo_ffi(cudaStream_t stream, ffi::Span<const int64_t> gdims, ffi::Sp
 template <ffi::DataType T>
 ffi::Error transpose_ffi(cudaStream_t stream, ffi::Span<const int64_t> gdims, ffi::Span<const int64_t> pdims,
                          int64_t transpose_comm_backend, int64_t halo_comm_backend, int64_t transpose_type,
-                         bool contiguous, ffi::Buffer<T> input, ffi::Buffer<ffi::DataType::U8> workspace,
+                         bool contiguous, ffi::Buffer<T> input, ffi::Buffer<ffi::DataType::S8> workspace,
                          ffi::Result<ffi::Buffer<T>> output) {
   if (gdims.size() != 3) { return ffi::Error::InvalidArgument("gdims must have 3 elements"); }
   if (pdims.size() != 2) { return ffi::Error::InvalidArgument("pdims must have 2 elements"); }
@@ -284,7 +284,7 @@ XLA_FFI_DEFINE_HANDLER_SYMBOL(pfft_C64, pfft3d_ffi<ffi::DataType::C64>,
                                   .Attr<bool>("contiguous")
                                   .Attr<int64_t>("decomposition")
                                   .Arg<ffi::Buffer<ffi::DataType::C64>>() // input
-                                  .Arg<ffi::Buffer<ffi::DataType::U8>>()  // workspace (bytes)
+                                  .Arg<ffi::Buffer<ffi::DataType::S8>>()  // workspace (bytes)
                                   .Ret<ffi::Buffer<ffi::DataType::C64>>() // output
 );
 
@@ -300,11 +300,11 @@ XLA_FFI_DEFINE_HANDLER_SYMBOL(pfft_C128, pfft3d_ffi<ffi::DataType::C128>,
[…]
