### 2026-02-07 64914c3
Update benchmarks and demo features
