### 2026-02-07 dd9a2ee
Update benchmark scripts and add zenodo badge
