### 2026-02-07 82e1e5a
Update CITATION (#110)
* Update benchmark scripts and add zenodo badge

* Update citation DOI
