### 2026-02-07 e274e6c
Bump version to 0.3.0 (#111)
