+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["scikit-build-core>=0.4.3", "nanobind>=1.3.2", "jax>=0.4.35"]
 build-backend = "scikit_build_core.build"
 [project]
 name = "jaxdecomp"
-version = "0.2.9"
+version = "0.3.0"
 description = "JAX bindings for the cuDecomp library"
 authors = [
   { name = "Wassim Kabalan" },
