### 2026-02-07 ef39894
Fix build wheel action (#112)
* Bump version to 0.3.0

* Fix scipy and numpy build issues in GitHub Actions
