+++ .github/workflows/github-deploy.yml
@@ -21,6 +21,7 @@ jobs:
         env:
          CIBW_BUILD: "cp310-manylinux_x86_64 cp311-manylinux_x86_64 cp312-manylinux_x86_64"
          CIBW_BUILD_VERBOSITY: 2
+         CIBW_ENVIRONMENT: "PIP_ONLY_BINARY=scipy,numpy"
       - uses: actions/upload-artifact@v4
         with:
           name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
