### 2026-02-07 3a22db1
Allow using vmap to do a nested sharding (batched pfft3d)
