+++ tests/conftest.py
@@ -104,6 +104,8 @@ def create_mesh(pdims, axis_type, transposed_devices=False):
 
     size = jax.device_count()
     assert len(pdims) == 2
+    if prod(pdims) > size:
+        raise ValueError(f'prod(pdims)={prod(pdims)} exceeds device count {size}')
     if prod(pdims) != size:
         devices_id = jax.devices()[: prod(pdims)]
     else:
@@ -127,14 +129,20 @@ def create_batched_mesh(batch_size, spatial_pdims, axis_type):
 
     size = jax.device_count()
     assert len(spatial_pdims) == 2
-    assert batch_size * prod(spatial_pdims) == size, 'batch_size * prod(spatial_pdims) must equal device count'
+    total = batch_size * prod(spatial_pdims)
+    if total > size:
+        raise ValueError(f'batch_size * prod(spatial_pdims)={total} exceeds device count {size}')
+    if total != size:
+        devices = jax.devices()[:total]
+    else:
+        devices = jax.devices()
 
     mesh_shape = (batch_size, *spatial_pdims)
     if axis_type == 'explicit':
         axis_types = (AxisType.Explicit,) * len(mesh_shape)
     else:
         axis_types = (AxisType.Auto,) * len(mesh_shape)
-    return jax.make_mesh(mesh_shape, ('c', 'z', 'y'), axis_types=axis_types)
+    return jax.make_mesh(mesh_shape, ('c', 'z', 'y'), devices=devices, axis_types=axis_types)
 
 
 def create_batched_spmd_array(global_shape, mesh):
@@ -143,6 +151,9 @@ def create_batched_spmd_array(global_shape, mesh):
     from jax.sharding import PartitionSpec as P
 
     assert len(global_shape) == 4
+    assert global_shape[0] % mesh.shape['c'] == 0, "Batch dim not divisible by mesh 'c'"
+    assert global_shape[1] % mesh.shape['z'] == 0, "Spatial dim not divisible by mesh 'z'"
+    assert global_shape[2] % mesh.shape['y'] == 0, "Spatial dim not divisible by mesh 'y'"
 
     local_shape = (
         global_shape[0] // mesh.shape['c'],
@@ -157,6 +168,26 @@ def create_batched_spmd_array(global_shape, mesh):
         sharding,
         data_callback=lambda x: jax.random.normal(jax.random.PRNGKey(process_slices(x)), local_shape),
     )
+
+    return global_array
+
+
+def create_batched_ones_spmd_array(global_shape, mesh):
+    import jax
+    from jax.sharding import NamedSharding
+    from jax.sharding import PartitionSpec as P
+
+    assert len(global_shape) == 4
+
+    local_shape = (
+        global_shape[0] // mesh.shape['c'],
+        global_shape[1] // mesh.shape['z'],
+        global_shape[2] // mesh.shape['y'],
+        global_shape[3],
+    )
+
+    sharding = NamedSharding(mesh, P('c', 'z', 'y'))
+    global_array = jax.make_array_from_callback(global_shape, sharding, data_callback=lambda _: jax.numpy.ones(local_shape))
 
     return global_array
 
