### 2026-02-09 65e5380
docs: add machine-specific installation notes to installation guide
