### 2026-02-09 929e2e3
Update docs/installation.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
