### 2026-02-09 9a56279
Support nested sharding (#113)
* Allow using vmap to do a nested sharding (batched pfft3d)

* Add notebooks on nested parallelism and tpu fft demo, rename demo_features notebook

* Final fix

* Fix typo in filename and add colab badge to nested parallelism notebook

* Update strong scaling script

* docs: add machine-specific installation notes to installation guide

* Update docs/installation.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update examples/03-nested-parallelism.ipynb

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

C
[…]
