+++ src/jaxdecomp/_src/jax/halo.py
@@ -378,8 +378,9 @@ def partition(
     pencil_type = get_pencil_type(mesh)
     spec_for_validation = input_sharding.spec
     if arg_infos[0].ndim == 4:
-        assert input_sharding.spec[0] is None
-        spec_for_validation = input_sharding.spec[1:]
+        if input_sharding.spec[0] is None:
+            spec_for_validation = input_sharding.spec[1:]
+        # else: sharded batch — full spec validates against full mesh
 
     validate_spec_matches_mesh(spec_for_validation, mesh)
 
@@ -422,10 +423,10 @@ def transpose_rule(cotangent: Array, x: Array, halo_extents: HaloExtentType, hal
 @spmd_halo_primitive.def_batching_rule
 def batching_rule(
     batched_args: tuple[Array],
-    batched_axis,
+    batched_axis: tuple[int | None, ...],
     halo_extents: HaloExtentType,
     halo_periods: Periodicity,
-) -> Array:
+) -> tuple[Array, int]:
     """
     Batching rule for the halo exchange operation.
 
@@ -433,8 +434,8 @@ def batching_rule(
     ----------
     batched_args : Tuple[Array]
         Batched input arrays.
-    batched_axis : int
-        Batch axis.
+    batched_axis : tuple[int | None, ...]
+        Batch axis for each operand.
     halo_extents : HaloExtentType
         Extents of the halo in X and Y directions.
     halo_periods : Periodicity
@@ -442,8 +443,8 @@ def batching_rule(
 
     Returns
     -------
-    Array
-        Resulting array after the halo exchange operation.
+    tuple[Array, int]
+        Resulting array and the output batch axis.
     """
     (x,) = batched_args
     (bd,) = batched_axis
