### 2026-02-09 #8852 @ASKabalan
Hello @MoraruMaxim,

Thank you for your patience. I have significantly updated the package to address previous issues and streamlined the installation process.

Here is a summary of the changes:

### 1. Core Updates & Installation

* **JAX Compatibility:** Updated to support JAX v0.9.0.
* **New FFI:** Updated the C++ binding API to utilize the new JAX FFI.
* **Documentation:** Comprehensive updates to the documentation.
* **Installation:** The build and install process is now much simpler.

### 2. Benchmarks

I have re-run the benchmarks with the updated setup. Below are the current results:

**FFT Performance:**

**IFFT Performance:**

**Weak Scaling (Fixed 64x128x128 per GPU):**

**Note on NVLINK:**
I discovered that `cuDecomp` has ![NVLINK disabled by default](https://nvidia.github.io/cuDecomp/env_vars.html#cudecomp-enable-cumem). I have just launched a new set of benchmarks with this variable activated to demonstrate full hardware capability. I expect those results to be ready in approximately 2 days.

### 3. Examples

I have added new notebooks to the repository. Specifically, if a reviewer has access to a multi-device TPU environment, the following example is now available:

* [02-tpu_fft_demo.ipynb](https://github.com/DifferentiableUniverseInitiative/jaxDecomp/blob/main/examples/02-tpu_fft_demo.ipynb)

I will post the updated NVLINK benchmark results as soon as they finish.

Please find attached the CSV and a way to plot them with my other package in the zipped folder

[PERFOLOGS.tar.gz](https://github.com/user-attachments/files/25187795/PERFOLOGS.tar.gz)
