### 2026-02-09 450f5da
Doc and isntallation updates (#114)
