+++ docs/installation.md
@@ -64,7 +64,7 @@ Then proceed with installing `jaxDecomp` with cuDecomp support.
 
 ### IDRIS [Jean Zay](http://www.idris.fr/eng/jean-zay/cpu/jean-zay-cpu-hw-eng.html) HPE SGI 8600 supercomputer
 
-As of February 2025, loading modules **in this exact order** works:
+As of February 2026, loading modules **in this exact order** works:
 
 ```bash
 module load nvidia-compilers/25.1 cuda/12.6.3 openmpi/4.1.6-cuda nccl/2.26.2-1-cuda cudnn  cmake
@@ -83,23 +83,6 @@ pip install git+https://github.com/DifferentiableUniverseInitiative/jaxDecomp -C
 > For example for A100 you need to load `module load arch/a100` first.
 > You also need to set the CXXFLAGS to `export CXXFLAGS="-tp=zen2 -noswitcherror"` if you are using the H100 or A100 partition or if you are using AMD CPUs in general.
 > More info in [Jean Zay documentation](http://www.idris.fr/eng/jean-zay/gpu/jean-zay-gpu-exec_partition_slurm-eng.html#a100_partition_gpu_p5).
-
-
-### NERSC [Perlmutter](https://docs.nersc.gov/systems/perlmutter/architecture/) HPE Cray EX supercomputer
-
-As of November 2022:
-
-```bash
-module load PrgEnv-nvhpc python
-export CRAY_ACCEL_TARGET=nvidia80
-
-# Install JAX
-pip install --upgrade "jax[cuda]"
-
-# Install jaxDecomp w/ cuDecomp
-export CMAKE_PREFIX_PATH=/opt/nvidia/hpc_sdk/Linux_x86_64/22.5/cmake
-pip install git+https://github.com/DifferentiableUniverseInitiative/jaxDecomp -Ccmake.define.JD_CUDECOMP_BACKEND=ON
-```
 
 ---
 
