### 2026-02-09 #8852 @ASKabalan
I updated installation here https://jaxdecomp.readthedocs.io/en/latest/installation.html#cudecomp-backend-advanced-hpc

It should be much easier to install now
