### 2026-02-11 341a143
update JaxTimer and jax hpc profiler dependencies
