### 2026-02-11 b2f41a0
update JaxTimer and jax hpc profiler dependencies (#115)
