### 2026-02-16 #8852 @MoraruMaxim
> I updated installation here https://jaxdecomp.readthedocs.io/en/latest/installation.html#cudecomp-backend-advanced-hpc
>
> It should be much easier to install now

Looks like it worked. I was able to install the `cuDecomp` backend. Could you please point me to the benchmarks I should run?

I tried to run the following benchmark: `https://github.com/DifferentiableUniverseInitiative/jaxDecomp/blob/main/benchmarks/slurm_strong_scaling.sh`

but run into the following error:
```
ImportError: cannot import name 'JaxTimer' from 'jax_hpc_profiler'
```
