### 2026-02-16 #8852 @ASKabalan
> but run into the following error:
>
> ```
> ImportError: cannot import name 'JaxTimer' from 'jax_hpc_profiler'
> ```

Do you have jax-hpc-profiler v0.3.2?

If not just install pip install `git+https://github.com/ASKabalan/jax-hpc-profiler.git`

Should work
