### 2026-02-16 #8852 @MoraruMaxim
> > but run into the following error:
> > ```
> > ImportError: cannot import name 'JaxTimer' from 'jax_hpc_profiler'
> > ```
>
> Do you have jax-hpc-profiler v0.3.2?
>
> If not just install pip install `git+https://github.com/ASKabalan/jax-hpc-profiler.git`
>
> Should work

This works! However, it looks like I built the CPU version by following the [new instructions](https://jaxdecomp.readthedocs.io/en/latest/installation.html#cudecomp-backend-advanced-hpc): ` An NVIDIA GPU may be present on this machine, but a CUDA-enabled jaxlib is not installed. Falling back to cpu.` Did I miss something?
