### 2026-02-16 #8852 @ASKabalan
This is a JAX issue

Did you install jax by doing `pip install jax[cuda]`

They recently did something with the init that made it very hard to find the correct NVHPC to use with jax

https://docs.jax.dev/en/latest/installation.html#nvidia-gpu

If you have an error saying cuSparse not found just install jax local

https://jaxdecomp.readthedocs.io/en/latest/installation.html#troubleshooting
