### 2026-02-16 #8852 @MoraruMaxim
I am confused with the [building instructions](https://jaxdecomp.readthedocs.io/en/latest/installation.html#cudecomp-backend-advanced-hpc).

In order to install the `cuDecomp Backend (Advanced / HPC)` backend do I still need to go through step 1?

```
pip install --upgrade "jax[cuda]"
# and
pip install jaxdecomp
```

nvhpc version: `nvhpc/25.1`
