### 2026-02-16 #8852 @ASKabalan
Yes both need JAX
JAX wouldn't allow lowering (or extending) primitive to GPU unless it has the GPU plugin itself

Also the NVHPC version is good
