### 2026-02-16 #8852 @MoraruMaxim
The installation step worked but now I have an error at runtime error:

```
ERROR:2026-02-16 14:25:32,221:jax._src.xla_bridge:491: Jax plugin configuration error: Exception when calling jax_plugins.xla_cuda12.initialize()
Traceback (most recent call last):
  File "<conda_path>/lib/python3.11/site-packages/jax_plugins/xla_cuda12/__init__.py", line 201, in _version_check
    version = get_version()
              ^^^^^^^^^^^^^
RuntimeError: jaxlib/cuda/versions_helpers.cc:72: operation cublasGetVersion( nullptr, &version) failed: cuBlas internal error

Traceback (most recent call last):
  File "<conda_path>/lib/python3.11/site-packages/jax/_src/xla_bridge.py", line 489, in discover_pjrt_plugins
    plugin_module.initialize()
  File "<conda_path>/lib/python3.11/site-packages/jax_plugins/xla_cuda12/__init__.py", line 328, in initialize
    _check_cuda_versions(raise_on_first_error=True)
  File "<conda_path>/lib/python3.11/site-packages/jax_plugins/xla_cuda12/__init__.py", line 261, in _check_cuda_versions
    cublas_version = _version_check("cuBLAS", cuda_versions.cublas_get_version,
                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "<conda_path>/lib/python3.11/site-packages/jax_plugins/xla_cuda12/__init__.py", line 205, in _version_check
    raise RuntimeError(err_msg) from e
```

Also, this is probably not the best place to debug it. Let me open an issue.
