### 2026-02-23 #8852 @MoraruMaxim
@diehlpk
Unfortunately, I was not able to successfully build jaxdecomp, despite multiple attempts on two different systems (IB + A100 and Slingshot + A100) using both a GPU-aware MPI implementation and NCCL.

I sincerely apologize, but since I am unable to reproduce the reported performance results, I feel I must withdraw from this review. I regret any inconvenience this may cause and appreciate the opportunity to have been considered as a reviewer.
