### 2026-02-23 #8852 @ASKabalan
Hello

I explained in details how to install here https://jaxdecomp.readthedocs.io/en/latest/installation.html#troubleshooting

The blocking issue for the other reviewer was that he did not have cudnn >=CUDNN 9.8 on his cluster
Which is blocked by JAX and sadly I cannot do anything about it.

https://docs.jax.dev/en/latest/installation.html#pip-installation-nvidia-gpu-cuda-installed-locally-harder

I am more than happy to help the new reviewers if they need any help
