### 2026-03-11 #8852 @K20shores
@diehlpk @ASKabalan I was on vacation for two weeks and am just now getting to this.

I was able to install the software on Derecho at the NCAR Wyoming Supercomputing Center. However, I've been having trouble getting all of the tests to run, which I'd like to get working before I try the benchmarks. I have to go through the installation process again since the available cuda and cudnn packages have changed on Derecho since I did the install. I'll do that tonight and report back where I am in the process.
