### 2026-03-12 #8852 @ASKabalan
@K20shores please feel free to share with me if you are encountering issues.
I have made significant changes to make the package work with `JAX 0.9.0` and the tests where passing for me
One thing to note is that if you run `pytest -v` with `cudecomp` test cases included you might encounter some issues because of the `pytest` scheduling sometimes doesn't work very smoothly with `cudecomp` handling of communicators and initialization this is why I added this [shell](https://github.com/DifferentiableUniverseInitiative/jaxDecomp/blob/main/tests/run_all_tests.sh)
