### 2026-03-17 #8852 @K20shores
@diehlpk @ASKabalan

This is where I am:

```
conda create --name jaxdecomp python=3.13
conda activate jaxdecomp
module load cuda/12.9.0 cudnn/9.8.0.87-12
pip install --upgrade "jax[cuda]"
```

Despite having cuda and cudnn loaded, some nvidia pacakges were installed:

```
(jaxdecomp) [kshores@deg0035] ~ $ pip list
Package                  Version
------------------------ -----------
certifi                  2025.4.26
cftime                   1.6.4.post1
colorlog                 6.9.0
contourpy                1.3.2
cycler                   0.12.1
fonttools                4.58.0
jax                      0.9.1
jax-cuda12-pjrt          0.9.1
jax-cuda12-plugin        0.9.1
jaxlib                   0.9.1
kiwisolver               1.4.8
matplotlib               3.10.3
ml_dtypes                0.5.4
mplcursors               0.6
netCDF4                  1.7.2
numpy                    2.4.3
nvidia-cublas-cu12       12.9.1.4
nvidia-cuda-cccl-cu12    12.9.27
nvidia-cuda-cupti-cu12   12.9.79
nvidia-cuda-nvcc-cu12    12.9.86
nvidia-cuda-nvrtc-cu12   12.9.86
nvidia-cuda-runtime-cu12 12.9.79
nvidia-cudnn-cu12        9.20.0.48
nvidia-cufft-cu12        11.4.1.4
nvidia-cusolver-cu12     11.7.5.82
nvidia-cusparse-cu12     12.5.10.65
nvidia-nccl-cu12         2.29.7
nvidia-nvjitlink-cu12    12.9.86
nvidia-nvshmem-cu12      3.5.21
opt_einsum               3.4.0
packaging                25.0
pandas                   2.2.3
pillow                   11.2.1
pip                      26.0.1
pyparsing                3.2.3
python-dateutil          2.9.0.post0
pytz                     2025.2
scipy                    1.17.1
six                      1.17.0
tqdm                     4.67.1
tzdata                   2025.2
xarray                   2025.4.0
```

i continued on anyways

```
pip install jaxdecomp
```

Then I tried to run the tests in your repository

```
(jaxdecomp) [kshores@deg0035] ~/jaxDecomp $ python --version
Python 3.13.12
(jaxdecomp) [kshores@deg0035] ~/jaxDecomp $ module list

Currently Loaded Modules:
  1) ncarenv/25.10 (S)   2) craype/2.7.34   3) intel/2025.2.1   4) ncarcompilers/1.1.0   5) libfabric/1.22.0   6) cray-mpich/8.1.32   7) hdf5/1.14.6   8) netcdf/4.9.3   9) conda/latest  10) cuda/12.9.0  11) cudnn/9.8.0.87-12

  Where:
   S:  Module is Sticky, requires --force to unload or purge

(jaxdecomp) [kshores@deg0035] ~/jaxDecomp $ pip list
Package                  Version
------------------------ -----------
certifi                  2025.4.26
cftime                   1.6.4.post1
colorlog                 6.9.0
contourpy                1.3.2
cycler                   0.12.1
fonttools                4.58.0
jax                      0.9.1
jax-cuda12-pjrt          0.9.1
jax-cuda12-plugin        0.9.1
jaxdecomp                0.3.0
jaxlib                   0.9.1
jaxtyping                0.3.9
kiwisolver               1.4.8
matplotlib               3.10.3
ml_dtypes                0.5.4
mplcursors               0.6
netCDF4                  1.7.2
numpy                    2.4.3
nvidia-cublas-cu12       12.9.1.4
nvidia-cuda-cccl-cu12    12.9.27
nvidia-cuda-cupti-cu12   12.9.79
nvidia-cuda-nvcc-cu12    12.9.86
nvidia-cuda-nvrtc-cu12   12.9.86
nvidia-cuda-runtime-cu12 12.9.79
nvidia-cudnn-cu12        9.20.0.48
nvidia-cufft-cu12        11.4.1.4
nvidia-cusolver-cu12     11.7.5.82
nvidia-cusparse-cu12     12.5.10.65
nvidia-nccl-cu12         2.29.7
nvidia-nvjitlink-cu12    12.9.86
nvidia-nvshmem-cu12      3.5.21
opt_einsum               3.4.0
packaging                25.0
pandas                   2.2.3
pillow                   11.2.1
pip                      26.0.1
pyparsing                3.2.3
python-dateutil          2.9.0.post0
pytz                     2025.2
scipy                    1.17.1
six                      1.17.0
tqdm                     4.67.1
tzdata                   2025.2
wadler_lindig            0.1.7
xarray                   2025.4.0

(jaxdecomp) [kshores@deg0035] ~/jaxDecomp $ git status
On branch main
Your branch is up to date with 'origin/main'.

nothing to commit, working tree clean
(jaxdecomp) [kshores@deg0035] ~/jaxDecomp $ pytest
-bash: pytest: command not found
(jaxdecomp) [kshores@deg0035] ~/jaxDecomp $ pip install 'jaxdecomp[test]'
Looking in indexes: https://pypi.org/simple, https://pip.repos.neuron.amazonaws.com, https://pypi.ngc.nvidia.com
Requirement already satisfied: jaxdecomp[test] in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (0.3.0)
Requirement already satisfied: jaxtyping>=0.2.0 in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jaxdecomp[test]) (0.3.9)
Requirement already satisfied: jax>=0.4.35 in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jaxdecomp[test]) (0.9.1)
Collecting pytest>=8.0.0 (from jaxdecomp[test])
  Downloading pytest-9.0.2-py3-none-any.whl.metadata (7.6 kB)
Collecting pytest-cov (from jaxdecomp[test])
  Downloading pytest_cov-7.0.0-py3-none-any.whl.metadata (31 kB)
Collecting pytest-mock (from jaxdecomp[test])
  Downloading pytest_mock-3.15.1-py3-none-any.whl.metadata (3.9 kB)
Collecting chex (from jaxdecomp[test])
  Downloading chex-0.1.91-py3-none-any.whl.metadata (18 kB)
Requirement already satisfied: jaxlib<=0.9.1,>=0.9.1 in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jax>=0.4.35->jaxdecomp[test]) (0.9.1)
Requirement already satisfied: ml_dtypes>=0.5.0 in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jax>=0.4.35->jaxdecomp[test]) (0.5.4)
Requirement already satisfied: numpy>=2.0 in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jax>=0.4.35->jaxdecomp[test]) (2.4.3)
Requirement already satisfied: opt_einsum in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jax>=0.4.35->jaxdecomp[test]) (3.4.0)
Requirement already satisfied: scipy>=1.13 in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jax>=0.4.35->jaxdecomp[test]) (1.17.1)
Requirement already satisfied: wadler-lindig>=0.1.3 in /glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages (from jaxtyping>=0.2.0->jaxdecomp[test]) (0.1.7)
Collecting iniconfig>=1.0.1 (from pytest>=8.0.0->jaxdecomp[test])
  Downloading iniconfig-2.3.0-py3-none-any.whl.metadata (2.5 kB)
Requirement already satisfied: packaging>=22 in /glade/u/home/kshores/.local/lib/python3.13/site-packages (from pytest>=8.0.0->jaxdecomp[test]) (25.0)
Collecting pluggy<2,>=1.5 (from pytest>=8.0.0->jaxdecomp[test])
  Downloading pluggy-1.6.0-py3-none-any.whl.metadata (4.8 kB)
Collecting pygments>=2.7.2 (from pytest>=8.0.0->jaxdecomp[test])
  Downloading pygments-2.19.2-py3-none-any.whl.metadata (2.5 kB)
Collecting absl-py>=2.3.1 (from chex->jaxdecomp[test])
  Downloading absl_py-2.4.0-py3-none-any.whl.metadata (3.3 kB)
Collecting typing_extensions>=4.15.0 (from chex->jaxdecomp[test])
  Downloading typing_extensions-4.15.0-py3-none-any.whl.metadata (3.3 kB)
Collecting toolz>=1.0.0 (from chex->jaxdecomp[test])
  Downloading toolz-1.1.0-py3-none-any.whl.metadata (5.1 kB)
Collecting coverage>=7.10.6 (from coverage[toml]>=7.10.6->pytest-cov->jaxdecomp[test])
  Downloading coverage-7.13.5-cp313-cp313-manylinux1_x86_64.manylinux_2_28_x86_64.manylinux_2_5_x86_64.whl.metadata (8.5 kB)
Downloading pytest-9.0.2-py3-none-any.whl (374 kB)
Downloading pluggy-1.6.0-py3-none-any.whl (20 kB)
Downloading iniconfig-2.3.0-py3-none-any.whl (7.5 kB)
Downloading pygments-2.19.2-py3-none-any.whl (1.2 MB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 1.2/1.2 MB 573.2 MB/s  0:00:00
Downloading chex-0.1.91-py3-none-any.whl (100 kB)
Downloading absl_py-2.4.0-py3-none-any.whl (135 kB)
Downloading toolz-1.1.0-py3-none-any.whl (58 kB)
Downloading typing_extensions-4.15.0-py3-none-any.whl (44 kB)
Downloading pytest_cov-7.0.0-py3-none-any.whl (22 kB)
Downloading coverage-7.13.5-cp313-cp313-manylinux1_x86_64.manylinux_2_28_x86_64.manylinux_2_5_x86_64.whl (253 kB)
Downloading pytest_mock-3.15.1-py3-none-any.whl (10 kB)
Installing collected packages: typing_extensions, toolz, pygments, pluggy, iniconfig, coverage, absl-py, pytest, pytest-mock, pytest-cov, chex
Successfully installed absl-py-2.4.0 chex-0.1.91 coverage-7.13.5 iniconfig-2.3.0 pluggy-1.6.0 pygments-2.19.2 pytest-9.0.2 pytest-cov-7.0.0 pytest-mock-3.15.1 toolz-1.1.0 typing_extensions-4.15.0

(jaxdecomp) [kshores@deg0035] ~/jaxDecomp $ pytest
========================================================================================================================================================================================================== test session starts ==========================================================================================================================================================================================================
rootdir: /glade/u/home/kshores/jaxDecomp
configfile: pyproject.toml
testpaths: tests
plugins: mock-3.15.1, jaxtyping-0.3.9, cov-7.0.0
collected 1109 items

tests/test_fft.py ssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss................................................................................................ssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss................................................sssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss [ 35%]
sssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.s.sssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss........................ssssssss.FFFFFFssssss                                                                                                                                                                                    [ 57%]
tests/test_halo.py ssssssssssssssssssssssssssssssss........................ssssssssFFFFFFssssss                                                                                                                                                                                                                                                                                                                                   [ 64%]
tests/test_transpose.py ssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss................................................................................................ssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss................................................ssssssssssssssssssssssssssssssssssssssssssssssssFFFFFFsss [ 99%]
sss                                                                                                                                                                                                                                                                                                                                                                                                                               [100%]

=============================================================================================================================================================================================================== FAILURES ================================================================================================================================================================================================================
_______________________________________________________________________________________________________________________________________________________________________________________________ test_sharded_vmap[auto-SLAB_XY-no_shardy] _______________________________________________________________________________________________________________________________________________________________________________________________

spatial_pdims = (4, 1), use_shardy = False, axis_type = 'auto'

    @pytest.mark.parametrize('use_shardy', use_shardy)
    @pytest.mark.parametrize('spatial_pdims', batched_decomp)
    def test_sharded_vmap(spatial_pdims, use_shardy, axis_type):
        if use_shardy and not ALLOW_SHARDY_PARTITIONER:
            pytest.skip(reason='Shardy partitioner is not supported in this JAX version use at least JAX 0.7.0')

        if axis_type == 'explicit':
            pytest.skip(reason='Explicit axis type not yet supported for sharded vmap tests')

        jax.config.update('jax_use_shardy_partitioner', use_shardy)

        batch_size = 2
        global_shape = (batch_size, 8, 8, 8)

        # Create 3D mesh: (batch, spatial_x, spatial_y)
        mesh = create_batched_mesh(batch_size, spatial_pdims, axis_type=axis_type)
        global_array = create_batched_spmd_array(global_shape, mesh)
        # Compute expected output sharding
>       out_sharding = jaxdecomp.get_fft_output_sharding(global_array.sharding)
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

tests/test_fft.py:523:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
/glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages/jaxdecomp/_src/pencil_utils.py:216: in get_fft_output_sharding
    pencil_type = get_pencil_type(mesh)
                  ^^^^^^^^^^^^^^^^^^^^^
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

mesh = Mesh(axis_sizes=(2, 4, 1), axis_names=('c', 'z', 'y'), axis_types=(Auto, Auto, Auto))

    def get_pencil_type(mesh: Mesh) -> Any:
        if mesh.empty:
            pdims = (1, 1)
        else:
            pdims = mesh.devices.shape
            if len(pdims) == 1:
                pdims = (1,) + pdims

            if len(pdims) != 2:
>               raise ValueError('Only one or two-dimensional meshes are supported.')
E               ValueError: Only one or two-dimensional meshes are supported.

/glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages/jaxdecomp/_src/pencil_utils.py:74: ValueError
________________________________________________________________________________________________________________________________________________________________________________________________ test_sharded_vmap[auto-SLAB_XY-shardy] _________________________________________________________________________________________________________________________________________________________________________________________________

spatial_pdims = (4, 1), use_shardy = True, axis_type = 'auto'

    @pytest.mark.parametrize('use_shardy', use_shardy)
    @pytest.mark.parametrize('spatial_pdims', batched_decomp)
    def test_sharded_vmap(spatial_pdims, use_shardy, axis_type):
        if use_shardy and not ALLOW_SHARDY_PARTITIONER:
            pytest.skip(reason='Shardy partitioner is not supported in this JAX version use at least JAX 0.7.0')

        if axis_type == 'explicit':
            pytest.skip(reason='Explicit axis type not yet supported for sharded vmap tests')

        jax.config.update('jax_use_shardy_partitioner', use_shardy)

        batch_size = 2
        global_shape = (batch_size, 8, 8, 8)

        # Create 3D mesh: (batch, spatial_x, spatial_y)
        mesh = create_batched_mesh(batch_size, spatial_pdims, axis_type=axis_type)
        global_array = create_batched_spmd_array(global_shape, mesh)
        # Compute expected output sharding
>       out_sharding = jaxdecomp.get_fft_output_sharding(global_array.sharding)
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

tests/test_fft.py:523:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
/glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages/jaxdecomp/_src/pencil_utils.py:216: in get_fft_output_sharding
    pencil_type = get_pencil_type(mesh)
                  ^^^^^^^^^^^^^^^^^^^^^
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

mesh = Mesh(axis_sizes=(2, 4, 1), axis_names=('c', 'z', 'y'), axis_types=(Auto, Auto, Auto))

    def get_pencil_type(mesh: Mesh) -> Any:
        if mesh.empty:
            pdims = (1, 1)
        else:
            pdims = mesh.devices.shape
            if len(pdims) == 1:
                pdims = (1,) + pdims

            if len(pdims) != 2:
>               raise ValueError('Only one or two-dimensional meshes are supported.')
E               ValueError: Only one or two-dimensional meshes are supported.

/glade/work/kshores/conda-envs/jaxdecomp/lib/python3.13/site-packages/jaxdecomp/_src/pencil_utils.py:74: ValueError
_______________________________________________________________________________________________________________________________________________________________________________________________ test_sharded_vmap[auto-SLAB_YZ-no_shardy] _______________________________________________________________________________________________________________________________________________________________________________________________

spatial_pdims = (1, 4), use_shardy = False, axis_type = 'auto'

    @pytest.mark.parametrize('use_shardy', use_shardy)
    @pytest.mark.parametrize('spatial_pdims', batched_decomp)
    def test_sharded_vmap(spatial_pdims, use_shardy, axis_type):
        if use_shardy and not ALLOW_SHARDY_PARTITIONER:
            pytest.skip(reason='Shardy partitioner is not supported in this JAX version use at least JAX 0.7.0')

        if axis_type == 'explicit':
            pytest.skip(reason='Explicit axis type not yet supported for sharded vmap tests')

        jax.config.update('jax_use_shardy_partitioner', use_shardy)

        batch_size = 2
        global_shape = (batch_size, 8, 8, 8)

        # Create 3D mesh: (batch, spatial_x, spatial_y)
        mesh = create_batched_mesh(batch_size, spatial_pdims, axis_type=axis_type)
        global_array = create_batched_spmd_array(global_shape, mesh)
        # Compute expected output sharding
>       out_sharding = jaxdecomp.get_fft_output_sharding(global_array.sharding)
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

tests/test_fft.py:523:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
[…]
