### 2026-03-18 #8852 @ASKabalan
Hello @K20shores ,

All of the tests are working if you install from github

`pip install git+https://github.com/DifferentiableUniverseInitiative/jaxDecomp`

I have not yet published a new PyPi version with the fixes

Also if you want to run the skipped tests you need to install the cuDecomp backend https://jaxdecomp.readthedocs.io/en/latest/installation.html#cudecomp-backend-advanced-hpc

then make sure to run the tests using this batch https://github.com/DifferentiableUniverseInitiative/jaxDecomp/blob/main/tests/run_all_tests.sh and not by doing pytest -v since the pytest scheduler might cause issues with how cuDecomp handles the communicators.

Please let me know if you need more help
