### 2026-03-23 #8852 @ASKabalan
@K20shores @diehlpk
I have remove the cudecomp plots from the paper and kept the JAX backend benchmarks only using only my code.
It should be much easier to reproduce
