### 2026-03-24 #8852 @K20shores
@ASKabalan I'm attempting to recreate the benchmarks. I translated the slurms script to pbs (with claude). I've been waiting in the queue all day. I'll let you know when I start to see results
