### 2026-04-06 #8852 @diehlpk
> [@ASKabalan](https://github.com/ASKabalan) I'm attempting to recreate the benchmarks. I translated the slurms script to pbs (with claude). I've been waiting in the queue all day. I'll let you know when I start to see results

@K20shores any luck to run the benchmarks?
