### 2026-04-20 #8852 @rafmudaf
~@ASKabalan Are there community guidelines somewhere?~
Sorry, now I see the Contributing.md :)

I'm looking at this point in the review form:
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

@ASKabalan How about the aspects of the "community guidelines" question beyond code formatting - are PRs generally welcome, do you expect an issue before a pull request? Also, how should users get in touch with questions? If it helps, here's one I put together. I'm sure I copied this from somewhere else, so feel free to adapt it, if it's helpful.
