### 2026-04-20 #8852 @ASKabalan
Hello @rafmudaf

Please tell me if [this addition](https://github.com/DifferentiableUniverseInitiative/jaxDecomp/blob/update-guidelines/CONTRIBUTING.md) is enough
