### 2026-04-24 #8852 @K20shores
@diehlpk I've obtained some benchmark results. I only have 32 GPUs available to me. Full disclosure, I had claude generate the plotting scripts for this data and I didn't look too closely because it's plotting

[jaxfft.csv](https://github.com/user-attachments/files/27061676/jaxfft.csv)
