### 2026-04-24 #8852 @K20shores
@diehlpk these seem to match the general trends. I think everything looks fine to me
