### 2026-05-14 #8852 @diehlpk
@ASKabalan Please have a look at https://github.com/DifferentiableUniverseInitiative/jaxDecomp/pull/120 and merge if you agree.
